% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{avatar}
\alias{avatar}
\alias{avatar_group}
\title{avatar}
\usage{
avatar(...)

avatar_group(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
The Avatar component is used to represent a user, and displays the profile picture, initials or fallback icon.
}
\details{
\enumerate{
\item Avatar Props
\itemize{
\item \bold{color}. Change avatar color. Type: NormalColors string. Default: default.
\item \bold{textColor}. Change avatar text color. Type: SimpleColors string. Default: default.
\item \bold{src}. Image source (local or remote). Type: string. Default: NA.
\item \bold{text}. Display text when image is missing. Type: string. Default: NA.
\item \bold{alt}. Display alt text when image is missing. Type: string. Default: NA.
\item \bold{size}. Avatar size. Type: NormalSizes number. Default: medium.
\item \bold{bordered}. Bordered avatar. Type: boolean. Default: false.
\item \bold{borderWeight}. Border weight for bordered avatar. Type: NormalWeights. Default: normal.
\item \bold{pointer}. Display pointer cursor on hover. Type: boolean. Default: false.
\item \bold{stacked}. Stacked display group. Type: boolean. Default: false.
\item \bold{zoomed}. Zoomed avatar. Type: boolean. Default: false.
\item \bold{squared}. Squared avatar. Type: boolean. Default: false.
\item \bold{icon}. Show icon in avatar. Type: ReactNode. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: span.
\item \bold{...}. Native props. Type: ImgHTMLAttributes. Default: NA.
}
\item Avatar Group Props
\itemize{
\item \bold{count}. Total count of avatars. Type: number. Default: NA.
\item \bold{animated}. Display translations animation on hover. Type: boolean. Default: true.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Native props. Type: HTMLAttributes. Default: NA.
}
}
}
\examples{
library(shiny)
library(shinyNextUI)

links <- c(
  "https://i.pravatar.cc/150?u=a042581f4e29026024d",
  "https://i.pravatar.cc/150?u=a042581f4e29026704d",
  "https://i.pravatar.cc/150?u=a04258114e29026702d",
  "https://i.pravatar.cc/150?u=a048581f4e29026701d",
  "https://i.pravatar.cc/150?u=a092581d4ef9026700d"
)

ui <- nextui_page(
  text("avatar()"),
  grid_container(
    gap = 2,
    grid(avatar(text = "Avatar")),
    grid(
      avatar(
        squared = TRUE,
        src = "https://i.pravatar.cc/150?u=a042581f4e29026024d"
      )
    ),
    grid(avatar(text = "Avatar", size = "xl")),
    grid(
      avatar(
        text = "Avatar",
        color = "secondary",
        textColor="white"
      )
    ),
    grid(
      avatar(
        src = "https://i.pravatar.cc/150?u=a042581f4e25056704b",
        color = "warning",
        bordered = TRUE
      )
    ),
    grid(
      avatar(
        src = "https://i.pravatar.cc/150?u=a042581f4e29026704d",
        zoomed = TRUE
      )
    )
  ),
  spacer(y = 2),
  text("avatar_group()"),
  grid_container(
    gap = 4,
    grid(
      xs = 12,
      avatar_group(
        count = 12,
        lapply(links, function(link) avatar(src = link))
      )
    )
  )
)

server <- function(input, output, session) {}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/avatar}.
}
