% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{checkboxgroup_input}
\alias{checkboxgroup_input}
\alias{update_checkboxgroup_input}
\title{Checkbox group input}
\usage{
checkboxgroup_input(inputId, ..., choices, selected = NULL)

update_checkboxgroup_input(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...,
  choices = NULL,
  selected = NULL
)
}
\arguments{
\item{inputId}{Unique input id.}

\item{...}{Props.}

\item{choices}{Radio choices.}

\item{selected}{Default selected value.}

\item{session}{Shiny session.}
}
\description{
Checkbox group input
}
