% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{collapse}
\alias{collapse}
\title{collapse}
\description{
Collapse display a list of high-level options that can expand/collapse to reveal more information.
}
\details{
\enumerate{
\item Collapse Props
\itemize{
\item \bold{expanded}. Manage the expand behaivor by prop. Type: boolean. Default: false.
\item \bold{title}. Collapse title content. Type: string React.ReactNode. Default: NA.
\item \bold{subtitle}. Collapse description content. Type: string React.ReactNode. Default: NA.
\item \bold{divider}. Show or hide the collapse divider. Type: boolean. Default: true.
\item \bold{disabled}. Disable collapse. Type: boolean. Default: false.
\item \bold{bordered}. Bordered collapse. Type: boolean. Default: false.
\item \bold{shadow}. Enable or disable the collapse shadow. Type: boolean. Default: false.
\item \bold{animated}. Enable or disable the collapse animation. Type: boolean. Default: true.
\item \bold{showArrow}. Show or hide the collapse right arrow. Type: boolean. Default: true.
\item \bold{arrowIcon}. Collapse arrow icon. Type: React.ReactNode. Default: NA.
\item \bold{contentLeft}. Left content for the collapse. Type: React.ReactNode. Default: NA.
\item \bold{borderWeight}. Border weight for bordered collapse. Type: NormalWeights. Default: light.
\item \bold{index}. Collapse's index, this value it's autogenerated by default. Type: number. Default: NA.
\item \bold{preventDefault}. Prevent default event on when is selected through the Space, Enter keys. Type: boolean. Default: true.
\item \bold{onChange}. Callback fired when the value is changed. Type: (e: React.ChangeEvent, index:number, value: boolean) => void. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Div native props. Type: HTMLAttributes. Default: NA.
}
\item Collapse Group Props
\itemize{
\item \bold{accordion}. Expand children collapse like an accordion. Type: boolean. Default: true.
\item \bold{animated}. Enable or disable the collapse group animation. Type: boolean. Default: true.
\item \bold{bordered}. Bordered collapse group. Type: boolean. Default: false.
\item \bold{splitted}. Separated collapse group. Type: boolean. Default: false.
\item \bold{shadow}. Enable or disable the collapse group shadow. Type: boolean. Default: false.
\item \bold{divider}. Show or hide the collapse group children divider. Type: boolean. Default: NA.
\item \bold{borderWeight}. Border weight for bordered collapse group. Type: NormalWeights. Default: light.
\item \bold{onChange}. Callback fired when a collapse children value is changed. Type: (index:number, value: boolean) => void. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Div native props. Type: HTMLAttributes. Default: NA.
}
}
}
\seealso{
See \url{https://nextui.org/docs/components/collapse}.
}
