% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/layout.R
\name{container}
\alias{container}
\title{container}
\usage{
container(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
A component for fixing an element's width to the current breakpoint.
}
\details{
\enumerate{
\item Container Props
\itemize{
\item \bold{fluid}. width: 100\% at all breakpoints. Type: boolean. Default: false.
\item \bold{responsive}. Sets a max-width at each responsive breakpoint. Type: boolean. Default: true.
\item \bold{xs}. Extra small devices (<650px). Type: boolean. Default: false.
\item \bold{sm}. Small devices (≥650px). Type: boolean. Default: false.
\item \bold{md}. Medium devices (≥960px). Type: boolean. Default: false.
\item \bold{lg}. Large devices (≥1280px). Type: boolean. Default: false.
\item \bold{xl}. Extra large devices (≥1400px). Type: boolean. Default: false.
\item \bold{gap}. Children's spacing. Type: number. Default: 2.
\item \bold{wrap}. CSS "flex-wrap". Type: Wrap. Default: wrap.
\item \bold{display}. CSS "display". Type: Display. Default: block.
\item \bold{justify}. CSS "justify-content". Type: Justify. Default: NA.
\item \bold{direction}. CSS "flex-direction". Type: Direction. Default: NA.
\item \bold{alignItems}. CSS "align-items". Type: AlignItems. Default: NA.
\item \bold{alignContent}. CSS "align-content". Type: AlignContent. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Native element props. Type: HTMLAttributes. Default: NA.
}
\item Row Props
\itemize{
\item \bold{fluid}. width: 100\%. Type: boolean. Default: true.
\item \bold{gap}. Children's spacing. Type: number. Default: 0.
\item \bold{wrap}. CSS "flex-wrap" property. Type: Wrap. Default: nowrap.
\item \bold{justify}. CSS "justify-content" property. Type: Justify. Default: flex-start.
\item \bold{align}. CSS "align-items" property. Type: AlignItems. Default: flex-start.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Native element props. Type: HTMLAttributes. Default: NA.
}
\item Col Props
\itemize{
\item \bold{span}. Column's width. Type: number. Default: 12.
\item \bold{offset}. Move columns to the right. Type: number. Default: 0.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Native element props. Type: HTMLAttributes. Default: NA.
}
}
}
\examples{

}
\seealso{
See \url{https://nextui.org/docs/layout/container}.
}
