% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createReactShinyInput}
\alias{createReactShinyInput}
\title{Create a reactR shiny input element}
\usage{
createReactShinyInput(
  inputId,
  class,
  default = NULL,
  configuration = list(),
  container = htmltools::tags$div
)
}
\arguments{
\item{inputId}{Unique input id.}

\item{class}{Element class. Must match the JavaScript class counterpart.}

\item{default}{Default value.}

\item{configuration}{Props.}

\item{container}{Default container.}
}
\value{
A list of tags.
}
\description{
This is used to create custom react element for R. Specifically
for radio and checkboxgroup which don't work with shiny.react.
}
