% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{custom_component}
\alias{custom_component}
\title{Documentation template for remaining components}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{motion}{Prop relation to the framer-motion react library.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Documentation template for remaining components
}
\keyword{internal}
