% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dropdow_menu}
\alias{dropdow_menu}
\alias{dropdown_item}
\alias{dropdown_section}
\alias{update_dropdown}
\title{Dropdown menu}
\usage{
dropdow_menu(inputId, ..., choices = NULL, selected = NULL)

dropdown_item(...)

dropdown_section(...)

update_dropdown(session = shiny::getDefaultReactiveDomain(), inputId, ...)
}
\arguments{
\item{inputId}{Unique input id.}

\item{...}{Props.}

\item{choices}{Radio choices.}

\item{selected}{Default selected value.}

\item{session}{Shiny session.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Dropdown menu
}
\note{
Container for related \link{dropdown_item}.
}
