% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/layout.R
\name{grid}
\alias{grid}
\alias{grid_container}
\title{grid}
\usage{
grid(...)

grid_container(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
The layout Grid adapts to screen size and orientation, ensuring consistency across layouts.
}
\details{
\enumerate{
\item Grid Props
\itemize{
\item \bold{xs}. extra small devices (<650px). Type: boolean. Default: false.
\item \bold{sm}. small devices (≥650px). Type: boolean. Default: false.
\item \bold{md}. medium devices (≥960px). Type: boolean. Default: false.
\item \bold{lg}. large devices (≥1280px). Type: boolean. Default: false.
\item \bold{xl}. extra large devices (≥1400px). Type: boolean. Default: false.
\item \bold{direction}. CSS "flex-direction" property. Type: Direction. Default: NA.
\item \bold{justify}. CSS "justify-content" property. Type: Justify. Default: NA.
\item \bold{alignItems}. CSS "align-items" property. Type: AlignItems. Default: NA.
\item \bold{alignContent}. CSS "align-content" property. Type: AlignContent. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Native element props. Type: HTMLAttributes. Default: NA.
}
\item Grid.Container Props
\itemize{
\item \bold{gap}. spacing of children. Type: number. Default: 0.
\item \bold{wrap}. CSS "flex-wrap" property. Type: Wrap. Default: wrap.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: div.
\item \bold{...}. Native element props. Type: HTMLAttributes. Default: NA.
}
}
}
\examples{
library(shiny)
library(shinyNextUI)

ui <- nextui_page(
  grid_container(
    gap = 2,
    justify = "center",
    grid(xs = 8, card(card_body(h1("Card 1")))),
    grid(xs = 4, card(card_body(h1("Card 2")))),
    grid(xs = 6, card(card_body(h1("Card 3")))),
    grid(xs = 6, card(card_body(h1("Card 4"))))
  )
)

server <- function(input, output, session) {}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/layout/grid}.
}
