% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/inputs.R
\name{input}
\alias{input}
\alias{text_input}
\alias{update_text_input}
\alias{numeric_input}
\alias{update_numeric_input}
\alias{date_input}
\alias{update_date_input}
\title{input}
\usage{
text_input(inputId, ..., value = default_value)

update_text_input(session = shiny::getDefaultReactiveDomain(), inputId, ...)

numeric_input(inputId, ..., value = default_value)

update_numeric_input(session = shiny::getDefaultReactiveDomain(), inputId, ...)

date_input(inputId, ..., value = default_value)

update_date_input(session = shiny::getDefaultReactiveDomain(), inputId, ...)
}
\arguments{
\item{inputId}{ID of the component.}

\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{value}{Starting value.}

\item{session}{Object passed as the \code{session} argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Input is a component that allows users to enter text. It can be used to get user inputs in forms, search fields, and more.
}
\details{
\itemize{
\item \bold{children}. Type: \code{ReactNode}. Default: NA.
\item \bold{variant}. Type: \code{flat | bordered | faded | underlined}. Default: "flat".
\item \bold{color}. Type: \code{default | primary | secondary | success | warning | danger}. Default: "default".
\item \bold{size}. Type: \code{sm | md | lg}. Default: "md".
\item \bold{radius}. Type: \code{none | sm | md | lg | full}. Default: NA.
\item \bold{label}. Type: \code{ReactNode}. Default: NA.
\item \bold{value}. Type: \code{string}. Default: NA.
\item \bold{defaultValue}. Type: \code{string}. Default: NA.
\item \bold{placeholder}. Type: \code{string}. Default: NA.
\item \bold{description}. Type: \code{ReactNode}. Default: NA.
\item \bold{errorMessage}. Type: \code{ReactNode | ((v: ValidationResult) => ReactNode)}. Default: NA.
\item \bold{validate}. Type: \code{(value: string) => ValidationError | true | null | undefined}. Default: NA.
\item \bold{validationBehavior}. Type: \code{native | aria}. Default: "native".
\item \bold{minLength}. Type: \code{number}. Default: NA.
\item \bold{maxLength}. Type: \code{number}. Default: NA.
\item \bold{pattern}. Type: \code{string}. Default: NA.
\item \bold{type}. Type: \code{text | email | url | password | tel | search | file}. Default: "text".
\item \bold{startContent}. Type: \code{ReactNode}. Default: NA.
\item \bold{endContent}. Type: \code{ReactNode}. Default: NA.
\item \bold{labelPlacement}. Type: \code{inside | outside | outside-left}. Default: "inside".
\item \bold{fullWidth}. Type: \code{boolean}. Default: true.
\item \bold{isClearable}. Type: \code{boolean}. Default: false.
\item \bold{isRequired}. Type: \code{boolean}. Default: false.
\item \bold{isReadOnly}. Type: \code{boolean}. Default: false.
\item \bold{isDisabled}. Type: \code{boolean}. Default: false.
\item \bold{isInvalid}. Type: \code{boolean}. Default: false.
\item \bold{baseRef}. Type: \code{RefObject<HTMLDivElement>}. Default: NA.
\item \bold{disableAnimation}. Type: \code{boolean}. Default: false.
\item \bold{classNames}. Type: \code{Partial<Record<'base' | 'label' | 'inputWrapper' | 'innerWrapper' | 'mainWrapper' | 'input' | 'clearButton' | 'helperWrapper' | 'description' | 'errorMessage', string>>}. Default: NA.
}
\itemize{
\item \bold{onChange}. Type: \code{React.ChangeEvent<HTMLInputElement>}. Default: NA.
\item \bold{onValueChange}. Type: \code{(value: string) => void}. Default: NA.
\item \bold{onClear}. Type: \code{() => void}. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)

ui <- nextui_page(
  div(
    class = "flex gap-1",
    text_input(
      inputId = "text",
      value = "Plop",
      placeholder = "Next UI",
      label = "Text input"
    ),
    textOutput("text_val")
  ),
  spacer(y = 5),
  div(
    class = "flex gap-1",
    numeric_input(
      inputId = "numeric",
      value = 10,
      label = "Numeric input"
    ),
    textOutput("numeric_val")
  ),
  spacer(y = 5),
  div(
    class = "flex gap-1",
    date_input(
      inputId = "date",
      value = "2023-12-11",
      label = "Date input"
    ),
    textOutput("date_val")
  )
)

server <- function(input, output, session) {
  output$text_val <- renderText(input$text)
  output$numeric_val <- renderText(input$numeric)
  output$date_val <- renderText(input$date)
}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/input}.
}
