% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{sizes}
\alias{sizes}
\alias{colors}
\alias{radiuses}
\alias{tabs_variants}
\title{Available sizes}
\format{
An object of class \code{character} of length 3.

An object of class \code{character} of length 6.

An object of class \code{character} of length 5.

An object of class \code{character} of length 4.
}
\usage{
sizes

colors

radiuses

tabs_variants
}
\description{
Available sizes

Available colors

Available radiuses

Available tabs variants
}
\keyword{datasets}
