% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poke_data}
\alias{poke_data}
\title{Pokemon API data}
\format{
\subsection{\code{poke_data}}{

A nested list with 151 entries. Each sublist contains:
\itemize{
\item name (char): Pokemon name.
\item description (char): Pokemon description.
\item shape (char): Pokemon shape.
\item sprites (list):
\itemize{
\item front_default (char): front sprite URL.
\item shiny_default (char): front sprite URL (shiny form).
}
\item ...
}
}
}
\source{
\url{https://pokeapi.co/docs/v2}
}
\usage{
poke_data
}
\description{
Extract of some data from the 151 first Pokemons.
}
\note{
Have a look to inst/app-doc/data-doc.html to get an
interactive overview.
}
\keyword{datasets}
