% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{run_app_test}
\alias{run_app_test}
\title{Helper to test package components}
\usage{
run_app_test(el, props = list(), children = tagList())
}
\arguments{
\item{el}{Component function like \link{card}.}

\item{props}{Component props. A list or leave NULL.}

\item{children}{Component children when applicable.
A list of tags or leave NULL.}
}
\value{
A shiny app object. Can be used with shinytest2.
}
\description{
Helper to test package components
}
\examples{
if (interactive()) {
 run_app_test(
  el = card,
  props = list(variant = "bordered"),
  children = tagList(
   card_header("Card title"),
   card_divider(),
   card_body(h1("Card body")),
   card_divider(),
   card_footer("Card Footer")
  )
 )
}
}
