% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{run_example}
\alias{run_example}
\title{Run shinyNextUI example}
\usage{
run_example(name)
}
\arguments{
\item{name}{Use \link{get_examples} to get
the available examples.}
}
\value{
Runs a Shiny app.
}
\description{
Run shinyNextUI example
}
