% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{snippet}
\alias{snippet}
\title{snippet}
\usage{
snippet(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Snippet is a component that can be used to display inline or multiline code snippets.
}
\details{
\itemize{
\item \bold{children}. Type: \code{ReactNode | ReactNode[]}. Default: NA.
\item \bold{size}. Type: \code{sm | md | lg}. Default: "md".
\item \bold{radius}. Type: \code{none | sm | md | lg}. Default: "lg".
\item \bold{symbol}. Type: \code{string | ReactNode}. Default: "$".
\item \bold{timeout}. Type: \code{number}. Default: "2000".
\item \bold{codeString}. Type: \code{string}. Default: NA.
\item \bold{tooltipProps}. Type: \code{TooltipProps}. Default: NA.
\item \bold{copyIcon}. Type: \code{ReactNode}. Default: NA.
\item \bold{checkIcon}. Type: \code{ReactNode}. Default: NA.
\item \bold{disableTooltip}. Type: \code{boolean}. Default: false.
\item \bold{disableCopy}. Type: \code{boolean}. Default: false.
\item \bold{hideCopyButton}. Type: \code{boolean}. Default: false.
\item \bold{hideSymbol}. Type: \code{boolean}. Default: false.
\item \bold{copyButtonProps}. Type: \code{ButtonProps}. Default: NA.
\item \bold{disableAnimation}. Type: \code{boolean}. Default: false.
\item \bold{classNames}. Type: \code{Partial<Record<'base' | 'content' | 'pre' | 'symbol' | 'copyButton' | 'checkIcon', string>>}. Default: NA.
}
\itemize{
\item \bold{onCopy}. Type: \code{(value: string | string[]) => void}. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  debug_react = TRUE,
  snippet("npm install @nextui-org/react", variant = "bordered"),
  spacer(y = 2),
  snippet("you can't copy me", hideCopyButton = TRUE),
  spacer(y = 2),
  snippet(
    color = "secondary",
    variant = "flat",
    span("npm install @nextui-org/react"),
    span("yarn add @nextui-org/react")
  )
)

server <- function(input, output, session) {

}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/snippet}.
}
