% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{.table}
\alias{.table}
\alias{table_header}
\alias{table_body}
\alias{table_col}
\alias{table_row}
\alias{table_cell}
\alias{table}
\title{Table widget}
\usage{
.table(...)

table_header(...)

table_body(...)

table_col(...)

table_row(...)

table_cell(...)

table(data = NULL, ...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{data}{Data to render.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Table widget
}
\keyword{internal}
