% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{text}
\alias{text}
\title{text}
\description{
Text component is the used to render text and paragraphs within an interface using well-defined typographic styles. It renders a \if{html}{\out{<p>}} tag by default.
}
\details{
\enumerate{
\item Text Props
\itemize{
\item \bold{h1 - h6}. Component name. Type: boolean. Default: false.
\item \bold{small}. Component name. Type: boolean. Default: false.
\item \bold{span}. Component name. Type: boolean. Default: false.
\item \bold{del}. Component name. Type: boolean. Default: false.
\item \bold{i}. Component name. Type: boolean. Default: false.
\item \bold{em}. Component name. Type: boolean. Default: false.
\item \bold{b}. Component name. Type: boolean. Default: false.
\item \bold{blockquote}. Component name. Type: boolean. Default: false.
\item \bold{transform}. text-transform prop. Type: TextTransforms. Default: none.
\item \bold{size}. Text size. Type: string number FontSize. Default: inherit.
\item \bold{color}. Text color. Type: SimpleColors string. Default: default.
\item \bold{weight}. Text weight. Type: TextWeights. Default: noset.
\item \bold{hideIn new}. Whether the Text should be hidden from selected breakpoint.. Type: xs sm md lg xl. Default: NA.
\item \bold{showIn new}. Whether the Text should be visible only from selected breakpoint.. Type: xs sm md lg xl. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: p.
}
}
}
\examples{
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  text(
    h1 = TRUE,
    size = 60,
    color = "primary",
    "Lets's",
    css = JS("{textGradient: '45deg, $blue600 -20\%, $pink600 50\%'}"),
    weight = "bold"
  ),
  text(
    h1 = TRUE,
    size = 60,
    "Make the Web",
    css = JS("{textGradient: '45deg, $purple600 -20\%, $pink600 100\%'}"),
    weight = "bold"
  ),
  text(
    h1 = TRUE,
    size = 60,
    "Prettier",
    css = JS("{textGradient: '45deg, $yellow600 -20\%, $red600 100\%'}"),
    weight = "bold"
  )
)

server <- function(input, output, session) {}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/text}.
}
