% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/inputs.R
\name{textarea}
\alias{textarea}
\alias{textarea_input}
\alias{update_textarea_input}
\title{textarea}
\usage{
textarea_input(inputId, ..., value = default_value)

update_textarea_input(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{inputId}{ID of the component.}

\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{value}{Starting value.}

\item{session}{Object passed as the \code{session} argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Textarea component is a multi-line Input which allows you to write large texts.
}
\details{
\itemize{
\item \bold{children}. Type: \code{ReactNode}. Default: NA.
\item \bold{minRows}. Type: \code{number}. Default: "3".
\item \bold{maxRows}. Type: \code{number}. Default: "8".
\item \bold{cacheMeasurements}. Type: \code{boolean}. Default: false.
\item \bold{variant}. Type: \code{flat | bordered | faded | underlined}. Default: "flat".
\item \bold{color}. Type: \code{default | primary | secondary | success | warning | danger}. Default: "default".
\item \bold{size}. Type: \code{sm | md | lg}. Default: "md".
\item \bold{radius}. Type: \code{none | sm | md | lg | full}. Default: NA.
\item \bold{label}. Type: \code{ReactNode}. Default: NA.
\item \bold{value}. Type: \code{string}. Default: NA.
\item \bold{defaultValue}. Type: \code{string}. Default: NA.
\item \bold{placeholder}. Type: \code{string}. Default: NA.
\item \bold{startContent}. Type: \code{ReactNode}. Default: NA.
\item \bold{endContent}. Type: \code{ReactNode}. Default: NA.
\item \bold{description}. Type: \code{ReactNode}. Default: NA.
\item \bold{errorMessage}. Type: \code{ReactNode | ((v: ValidationResult) => ReactNode)}. Default: NA.
\item \bold{validate}. Type: \code{(value: string) => ValidationError | true | null | undefined}. Default: NA.
\item \bold{validationBehavior}. Type: \code{native | aria}. Default: "native".
\item \bold{labelPlacement}. Type: \code{inside | outside | outside-left}. Default: "inside".
\item \bold{fullWidth}. Type: \code{boolean}. Default: true.
\item \bold{isRequired}. Type: \code{boolean}. Default: false.
\item \bold{isReadOnly}. Type: \code{boolean}. Default: NA.
\item \bold{isDisabled}. Type: \code{boolean}. Default: false.
\item \bold{isClearable}. Type: \code{boolean}. Default: false.
\item \bold{isInvalid}. Type: \code{boolean}. Default: false.
\item \bold{validationState}. Type: \code{valid | invalid}. Default: NA.
\item \bold{disableAutosize}. Type: \code{boolean}. Default: false.
\item \bold{disableAnimation}. Type: \code{boolean}. Default: false.
\item \bold{classNames}. Type: \code{Partial<Record<"base" | "label" | "inputWrapper" | "innerWrapper" | "input" | "description" | "errorMessage", string>>}. Default: NA.
}
\itemize{
\item \bold{onChange}. Type: \code{React.ChangeEvent<HTMLInputElement>}. Default: NA.
\item \bold{onValueChange}. Type: \code{(value: string) => void}. Default: NA.
\item \bold{onClear}. Type: \code{() => void}. Default: NA.
\item \bold{onHeightChange}. Type: \code{(height: number, meta: { rowHeight: number }) => void}. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)

ui <- nextui_page(
  div(
    class = "flex gap-5",
    action_button("update_text", "Update text"),
    textarea_input(
      inputId = "textarea",
      placeholder = "Enter your amazing ideas.",
      label = "Text area input",
      bordered = TRUE,
      color = "secondary",
      status = "secondary",
      helperColor = "error",
      helperText = "Click on update text"
    )
  ),
  textOutput("textarea_val")
)

server <- function(input, output, session) {
  output$textarea_val <- renderText(input$textarea)

  observeEvent(input$update_text, {
    update_textarea_input(
      inputId = "textarea",
      value = "Updated value"
    )
  })
}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/textarea}.
}
