% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{user}
\alias{user}
\alias{user_link}
\title{user}
\usage{
user(...)

user_link(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Flexible User Profile Component.
}
\details{
\enumerate{
\item User Props
\itemize{
\item \bold{color}. Change User Profile border color. Type: NormalColors string. Default: default.
\item \bold{src}. Image source (local or remote). Type: string. Default: NA.
\item \bold{name}. Display Text Next to Image. Type: string. Default: NA.
\item \bold{text}. Display Text when image is missing. Type: string. Default: NA.
\item \bold{alt}. Display alt text when image is missing. Type: string. Default: NA.
\item \bold{size}. User Profile size. Type: NormalSizes number. Default: medium.
\item \bold{bordered}. Bordered User Profile. Type: boolean. Default: false.
\item \bold{zoomed}. Zoomed User Profile. Type: boolean. Default: false.
\item \bold{pointer}. Display pointer cursor on hover. Type: boolean. Default: false.
\item \bold{squared}. Squared User Profile. Type: boolean. Default: false.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: span.
\item \bold{...}. Native props. Type: ImgHTMLAttributes. Default: NA.
}
}
}
\examples{
library(shiny)
library(shinyNextUI)

user <- shinyNextUI:::component("User")
colors <- c(
  "primary",
  "secondary",
  "success",
  "warning",
  "error"
)

bool_parm <- c(rep(TRUE, 3), rep(FALSE, 2))

user_config <- data.frame(
  src = c(
    "https://i.pravatar.cc/150?u=a042581f4e29026704d",
    "https://i.pravatar.cc/150?u=a04258114e29026702d",
    "https://i.pravatar.cc/150?u=a048581f4e29026701d",
    "https://i.pravatar.cc/150?u=a092581d4ef9026700d",
    "https://i.pravatar.cc/150?u=a042581f4e29026024d"
  ),
  name = c(
    "Ariana Wattson",
    "Jane Fisher",
    "William Howard",
    "Kristen Copper",
    "Tony Reichert"
  ),
  size = c("xs", "sm", "md", "lg", "xl"),
  color = colors,
  bordered = bool_parm,
  zoomed = bool_parm,
  pointer = bool_parm,
  squared = bool_parm,
  description = c(rep(NA, 2), rep("Description", 3))
)

user_factory <- function(src, name, size, color, bordered, zoomed, pointer, squared, description) {
  grid(
    user(
      src = src,
      name = name,
      size = size,
      color = color,
      bordered = bordered,
      zoomed = zoomed,
      pointer = pointer,
      squared = squared,
      description = description,
      user_link(href = "https://nextui.org/", "Link")
    )
  )
}

users <- purrr::pmap(user_config, user_factory)

ui <- nextui_page(
  text("user"),
  grid_container(
    gap = 2,
    xs = 12,
    sm = 6,
    users
  )
)

server <- function(input, output, session) {}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/user}.
}
