% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/inputs.R
\name{accordion}
\alias{accordion}
\alias{accordion_item}
\alias{update_accordion}
\title{accordion}
\usage{
accordion(inputId, ..., value = default_value)

accordion_item(...)

update_accordion(session = shiny::getDefaultReactiveDomain(), inputId, ...)
}
\arguments{
\item{inputId}{ID of the component.}

\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{value}{Starting value.}

\item{session}{Object passed as the \code{session} argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Accordion display a list of high-level options that can expand/collapse to reveal more information.

Accordion display a list of high-level options that can expand/collapse to reveal more information.
}
\details{
\itemize{
\item \bold{children}. The contents of the collection. Usually the array of AccordionItem. Type: ReactNode | ReactNode[]. Default: .
\item \bold{variant}. The accordion appearance style.. Type: light | shadow | bordered | splitted. Default: light.
\item \bold{selectionMode}. The type of selection that is allowed in the collection.. Type: none | single | multiple. Default: .
\item \bold{selectionBehavior}. The accordion selection behavior.. Type: toggle | replace. Default: toggle.
\item \bold{isCompact}. Whether all Accordion items should be smaller.. Type: boolean. Default: false.
\item \bold{isDisabled}. Whether the Accordion items are disabled.. Type: boolean. Default: .
\item \bold{showDivider}. WWhether to display a divider at the bottom of the each accordion item.. Type: boolean. Default: true.
\item \bold{DividerProps}. The divider component props.. Type: DividerProps. Default: NA.
\item \bold{hideIndicator}. Whether the Accordion items indicator is hidden.. Type: boolean. Default: .
\item \bold{disableAnimation}. Whether the Accordion items open/close animation is disabled.. Type: boolean. Default: .
\item \bold{disableIndicatorAnimation}. Whether the Accordion items indicator animation is disabled.. Type: boolean. Default: .
\item \bold{disallowEmptySelection}. Whether the collection allows empty selection.. Type: boolean. Default: .
\item \bold{keepContentMounted}. Whether the Accordion items content should be always mounted.. Type: boolean. Default: false.
\item \bold{fullWidth}. Whether the accordion should take up the full width of its parent container.. Type: boolean. Default: true.
\item \bold{motionProps}. The motion properties of the Accordion.. Type: MotionProps. Default: .
\item \bold{disabledKeys}. The item keys that are disabled. These items cannot be selected, focused, or otherwise interacted with.. Type: React.Key[]. Default: .
\item \bold{itemClasses}. The accordion items classNames.. Type: Classnames. Default: .
\item \bold{selectedKeys}. The currently selected keys in the collection (controlled).. Type: all | React.Key[]. Default: .
\item \bold{defaultSelectedKeys}. The initial selected keys in the collection (uncontrolled).. Type: all | React.Key[]. Default: .
\item \bold{disabledKeys}. The currently disabled keys in the collection (controlled).. Type: React.Key[]. Default: .
}
\itemize{
\item \bold{onSelectionChange}. Handler that is called when the selection changes.. Type: (keys: "all" ｜ Set<React.Key>) => any. Default: NA.
}
\itemize{
\item \bold{children}. The content of the component.. Type: ReactNode | string. Default: .
\item \bold{title}. The accordion item title.. Type: ReactNode | string. Default: .
\item \bold{subtitle}. The accordion item subtitle.. Type: ReactNode | string. Default: .
\item \bold{indicator}. The accordion item expanded indicator, usually an arrow icon.. Type: IndicatorProps. Default: .
\item \bold{startContent}. The accordion item start content, usually an icon or avatar.. Type: ReactNode. Default: .
\item \bold{motionProps}. The props to modify the framer motion animation. Use the variants API to create your own animation.. Type: MotionProps. Default: .
\item \bold{isCompact}. Whether the AccordionItem is compact.. Type: boolean. Default: false.
\item \bold{isDisabled}. The current disabled status.. Type: boolean. Default: false.
\item \bold{keepContentMounted}. Whether the AccordionItem content is kept mounted when closed.. Type: boolean. Default: false.
\item \bold{hideIndicator}. Whether the AccordionItem indicator is hidden.. Type: boolean. Default: false.
\item \bold{disableAnimation}. Whether the AccordionItem animation is disabled.. Type: boolean. Default: false.
\item \bold{disableIndicatorAnimation}. Whether the AccordionItem indicator animation is disabled.. Type: boolean. Default: false.
\item \bold{classNames}. Allows to set custom class names for the accordion item slots.. Type: Classnames. Default: NA.
}
\itemize{
\item \bold{onFocus}. Handler that is called when the element receives focus.. Type: (e: FocusEvent) => void. Default: NA.
\item \bold{onBlur}. Handler that is called when the element loses focus.. Type: (e: FocusEvent) => void. Default: NA.
\item \bold{onFocusChange}. Handler that is called when the element's focus status changes.. Type: (isFocused: boolean) => void. Default: NA.
\item \bold{onKeyDown}. Handler that is called when a key is pressed.. Type: (e: KeyboardEvent) => void. Default: NA.
\item \bold{onKeyUp}. Handler that is called when a key is released.. Type: (e: KeyboardEvent) => void. Default: NA.
\item \bold{onPress}. Handler called when the press is released over the target.. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onPressStart}. Handler called when a press interaction starts.. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onPressEnd}. Handler called when a press interaction ends, either over the target or when the pointer leaves the target.. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onPressChange}. Handler called when the press state changes.. Type: (isPressed: boolean) => void. Default: NA.
\item \bold{onPressUp}. Handler called when a press is released over the target, regardless of whether it started on the target or not.. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onClick}. The native button click event handler (Deprecated) use onPress instead.. Type: MouseEventHandler. Default: NA.
}

\itemize{
\item \bold{children}. Type: ReactNode | ReactNode[]. Default: .
\item \bold{variant}. Type: light | shadow | bordered | splitted. Default: "light".
\item \bold{selectionMode}. Type: none | single | multiple. Default: .
\item \bold{selectionBehavior}. Type: toggle | replace. Default: "toggle".
\item \bold{isCompact}. Type: boolean. Default: false.
\item \bold{isDisabled}. Type: boolean. Default: false.
\item \bold{showDivider}. Type: boolean. Default: true.
\item \bold{dividerProps}. Type: DividerProps. Default: .
\item \bold{hideIndicator}. Type: boolean. Default: false.
\item \bold{disableAnimation}. Type: boolean. Default: false.
\item \bold{disableIndicatorAnimation}. Type: boolean. Default: false.
\item \bold{disallowEmptySelection}. Type: boolean. Default: false.
\item \bold{keepContentMounted}. Type: boolean. Default: false.
\item \bold{fullWidth}. Type: boolean. Default: true.
\item \bold{motionProps}. Type: MotionProps. Default: .
\item \bold{disabledKeys}. Type: React.Key[]. Default: .
\item \bold{itemClasses}. Type: AccordionItemClassnames. Default: .
\item \bold{selectedKeys}. Type: all | React.Key[]. Default: .
\item \bold{defaultSelectedKeys}. Type: all | React.Key[]. Default: .
}
\itemize{
\item \bold{onSelectionChange}. Type: (keys: "all" | Set<React.Key>) => any. Default: NA.
}
\itemize{
\item \bold{children}. Type: ReactNode. Default: .
\item \bold{title}. Type: ReactNode. Default: .
\item \bold{subtitle}. Type: ReactNode. Default: .
\item \bold{indicator}. Type: IndicatorProps. Default: .
\item \bold{startContent}. Type: ReactNode. Default: .
\item \bold{motionProps}. Type: MotionProps. Default: .
\item \bold{isCompact}. Type: boolean. Default: false.
\item \bold{isDisabled}. Type: boolean. Default: false.
\item \bold{keepContentMounted}. Type: boolean. Default: false.
\item \bold{hideIndicator}. Type: boolean. Default: false.
\item \bold{disableAnimation}. Type: boolean. Default: false.
\item \bold{disableIndicatorAnimation}. Type: boolean. Default: false.
\item \bold{HeadingComponent}. Type: React.ElementType. Default: "h2".
\item \bold{classNames}. Type: AccordionItemClassnames. Default: .
}
\itemize{
\item \bold{onFocus}. Type: (e: FocusEvent) => void. Default: NA.
\item \bold{onBlur}. Type: (e: FocusEvent) => void. Default: NA.
\item \bold{onFocusChange}. Type: (isFocused: boolean) => void. Default: NA.
\item \bold{onKeyDown}. Type: (e: KeyboardEvent) => void. Default: NA.
\item \bold{onKeyUp}. Type: (e: KeyboardEvent) => void. Default: NA.
\item \bold{onPress}. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onPressStart}. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onPressEnd}. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onPressChange}. Type: (isPressed: boolean) => void. Default: NA.
\item \bold{onPressUp}. Type: (e: PressEvent) => void. Default: NA.
\item \bold{onClick}. Type: MouseEventHandler. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  debug_react = TRUE,
  p(class = "font-extrabold text-2xl uppercase", "accordion()"),
  accordion(
    inputId = "accordion1",
    value = "2",
    accordion_item(
      startContent = avatar(
        isBordered = TRUE,
        color = "primary",
        radius = "lg",
        src = "https://i.pravatar.cc/150?u=a042581f4e29026024d"
      ),
      onPress = JS("(e) => alert('Your pressed me')"),
      "plop",
      title = "Element 1",
      key = "1",
      subtitle = "subtitle"
    ),
    accordion_item("plop", title = "Element 2", key = "2")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "shadow variant accordion()"),
  accordion(
    inputId = "accordion2",
    isCompact = TRUE,
    variant = "shadow",
    value = "3",
    accordion_item("plop", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop", title = "Element 2", key = "2"),
    accordion_item("plop", title = "Element 3", key = "3")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "bordered variant accordion()"),
  accordion(
    inputId = "accordion3",
    isCompact = TRUE,
    variant = "bordered",
    accordion_item("plop", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop", title = "Element 2", key = "2")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "splitted variant accordion()"),
  accordion(
    inputId = "accordion4",
    isCompact = TRUE,
    variant = "splitted",
    motionProps= JS(
      '{
      variants: {
        enter: {
          y: 0,
          opacity: 1,
          height: "auto",
          transition: {
            height: {
              type: "spring",
              stiffness: 500,
              damping: 30,
              duration: 1,
            },
            opacity: {
              easings: "ease",
              duration: 1,
            },
          },
        },
        exit: {
          y: -10,
          opacity: 0,
          height: 0,
          transition: {
            height: {
              easings: "ease",
              duration: 0.25,
            },
            opacity: {
              easings: "ease",
              duration: 0.3,
            },
          },
        },
      },
    }'),
    value = "2",
    accordion_item("plop", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop", title = "Element 2", key = "2")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "Update accordion"),
  select_input(
    "select",
    label = "Accordion to open:",
    select_item(key = 1, value = "1", "1"),
    select_item(key = 2, value = "2", "2")
  ),
  spacer(y = 2),
  accordion(
    inputId = "accordion5",
    isCompact = TRUE,
    variant = "bordered",
    accordion_item("plop1", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop2", title = "Element 2", key = "2")
  )
)

server <- function(input, output, session) {
  observeEvent(input$select, {
    update_accordion(
      session,
      "accordion5",
      value = input$select
    )
  })
  observe(
    print(input$accordion1)
  )
}

if (interactive() || is_testing()) shinyApp(ui, server)
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  debug_react = TRUE,
  p(class = "font-extrabold text-2xl uppercase", "accordion()"),
  accordion(
    inputId = "accordion1",
    value = "2",
    accordion_item(
      startContent = avatar(
        isBordered = TRUE,
        color = "primary",
        radius = "lg",
        src = "https://i.pravatar.cc/150?u=a042581f4e29026024d"
      ),
      onPress = JS("(e) => alert('Your pressed me')"),
      "plop",
      title = "Element 1",
      key = "1",
      subtitle = "subtitle"
    ),
    accordion_item("plop", title = "Element 2", key = "2")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "shadow variant accordion()"),
  accordion(
    inputId = "accordion2",
    isCompact = TRUE,
    variant = "shadow",
    value = "3",
    accordion_item("plop", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop", title = "Element 2", key = "2"),
    accordion_item("plop", title = "Element 3", key = "3")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "bordered variant accordion()"),
  accordion(
    inputId = "accordion3",
    isCompact = TRUE,
    variant = "bordered",
    accordion_item("plop", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop", title = "Element 2", key = "2")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "splitted variant accordion()"),
  accordion(
    inputId = "accordion4",
    isCompact = TRUE,
    variant = "splitted",
    motionProps= JS(
      '{
      variants: {
        enter: {
          y: 0,
          opacity: 1,
          height: "auto",
          transition: {
            height: {
              type: "spring",
              stiffness: 500,
              damping: 30,
              duration: 1,
            },
            opacity: {
              easings: "ease",
              duration: 1,
            },
          },
        },
        exit: {
          y: -10,
          opacity: 0,
          height: 0,
          transition: {
            height: {
              easings: "ease",
              duration: 0.25,
            },
            opacity: {
              easings: "ease",
              duration: 0.3,
            },
          },
        },
      },
    }'),
    value = "2",
    accordion_item("plop", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop", title = "Element 2", key = "2")
  ),
  spacer(y = 2),
  p(class = "font-extrabold text-2xl uppercase", "Update accordion"),
  select_input(
    "select",
    label = "Accordion to open:",
    select_item(key = 1, value = "1", "1"),
    select_item(key = 2, value = "2", "2")
  ),
  spacer(y = 2),
  accordion(
    inputId = "accordion5",
    isCompact = TRUE,
    variant = "bordered",
    accordion_item("plop1", title = "Element 1", key = "1", subtitle = "subtitle"),
    accordion_item("plop2", title = "Element 2", key = "2")
  )
)

server <- function(input, output, session) {
  observeEvent(input$select, {
    update_accordion(
      session,
      "accordion5",
      value = input$select
    )
  })
  observe(
    print(input$accordion1)
  )
}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/accordion}.

See \url{https://nextui.org/docs/components/accordion}.
}
