% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{card}
\alias{card}
\alias{card_body}
\alias{card_header}
\alias{card_footer}
\title{card}
\usage{
card(...)

card_body(...)

card_header(...)

card_footer(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Card is a container for text, photos, and actions in the context of a single subject.
}
\details{
\itemize{
\item \bold{children}. Type: \code{ReactNode | ReactNode[]}. Default: NA.
\item \bold{shadow}. Type: \code{none | sm | md | lg}. Default: "md".
\item \bold{radius}. Type: \code{none | sm | md | lg}. Default: "lg".
\item \bold{fullWidth}. Type: \code{boolean}. Default: false.
\item \bold{isHoverable}. Type: \code{boolean}. Default: false.
\item \bold{isPressable}. Type: \code{boolean}. Default: false.
\item \bold{isBlurred}. Type: \code{boolean}. Default: false.
\item \bold{isFooterBlurred}. Type: \code{boolean}. Default: false.
\item \bold{isDisabled}. Type: \code{boolean}. Default: false.
\item \bold{disableAnimation}. Type: \code{boolean}. Default: false.
\item \bold{disableRipple}. Type: \code{boolean}. Default: false.
\item \bold{allowTextSelectionOnPress}. Type: \code{boolean}. Default: false.
\item \bold{classNames}. Type: \code{Partial<Record<'base' | 'header' | 'body' | 'footer', string>>}. Default: NA.
}
\itemize{
\item \bold{onPress}. Type: \code{(e: PressEvent) => void}. Default: NA.
\item \bold{onPressStart}. Type: \code{(e: PressEvent) => void}. Default: NA.
\item \bold{onPressEnd}. Type: \code{(e: PressEvent) => void}. Default: NA.
\item \bold{onPressChange}. Type: \code{(isPressed: boolean) => void}. Default: NA.
\item \bold{onPressUp}. Type: \code{(e: PressEvent) => void}. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  div(
    class = "grid gap-4 grid-cols-3 grid-rows-3 m-5",
    card(card_body("Simple card without anything")),
    card(
      variant = "bordered",
      card_header("Card title"),
      divider(),
      card_body(h1("Card body")),
      divider(),
      card_footer("Card Footer")
    ),
    card(
      #isBlurred = TRUE,
      isPressable = TRUE,
      onPress = JS("() => alert('You pressed me')"),
      shadow = "sm",
      className = "border-none bg-background/60 dark:bg-default-100/50 max-w-[610px]",
      card_body("Press me!")
    )
  )
)

server <- function(input, output, session) {

}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/card}.
}
