% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/inputs.R
\name{checkbox}
\alias{checkbox}
\alias{checkbox_input}
\alias{update_checkbox_input}
\title{checkbox}
\usage{
checkbox_input(inputId, ..., value = default_value)

update_checkbox_input(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{inputId}{ID of the component.}

\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{value}{Starting value.}

\item{session}{Object passed as the \code{session} argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Checkboxes allow users to select multiple items from a list of individual items, or to mark one individual item as selected.
}
\details{
\itemize{
\item \bold{children}. Type: \code{ReactNode}. Default: NA.
\item \bold{icon}. Type: \code{CheckboxIconProps}. Default: NA.
\item \bold{value}. Type: \code{string}. Default: NA.
\item \bold{name}. Type: \code{string}. Default: NA.
\item \bold{size}. Type: \code{sm | md | lg}. Default: "md".
\item \bold{color}. Type: \code{default | primary | secondary | success | warning | danger}. Default: "primary".
\item \bold{radius}. Type: \code{none | sm | md | lg | full}. Default: NA.
\item \bold{lineThrough}. Type: \code{boolean}. Default: false.
\item \bold{isSelected}. Type: \code{boolean}. Default: NA.
\item \bold{defaultSelected}. Type: \code{boolean}. Default: NA.
\item \bold{isRequired}. Type: \code{boolean}. Default: false.
\item \bold{isReadOnly}. Type: \code{boolean}. Default: NA.
\item \bold{isDisabled}. Type: \code{boolean}. Default: false.
\item \bold{isIndeterminate}. Type: \code{boolean}. Default: NA.
\item \bold{isInvalid}. Type: \code{boolean}. Default: false.
\item \bold{validationState}. Type: \code{valid | invalid}. Default: NA.
\item \bold{disableAnimation}. Type: \code{boolean}. Default: false.
\item \bold{classNames}. Type: \code{Partial<Record<"base"｜ "wrapper"｜ "icon"｜ "label", string>>}. Default: NA.
}
\itemize{
\item \bold{onChange}. Type: \code{React.ChangeEvent<HTMLInputElement>}. Default: NA.
\item \bold{onValueChange}. Type: \code{(isSelected: boolean) => void}. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)

ui <- nextui_page(
  div(
    class = "flex gap-1",
    action_button("update", "Toggle checkbox"),
    spacer(y = 2),
    checkbox_input(
      inputId = "checkbox",
      value = TRUE,
      "My checkbox",
      isRounded = TRUE,
      color = "warning",
      lineThrough = TRUE
    ),
    textOutput("check_val")
  )
)

server <- function(input, output, session) {
  output$check_val <- renderText(input$checkbox)
  observeEvent(input$update, {
    update_checkbox_input(session, "checkbox", value = !input$checkbox)
  })
}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/checkbox}.
}
