% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{image}
\alias{image}
\title{image}
\usage{
image(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
The Image component is used to display images with support for fallback.
}
\details{
\itemize{
\item \bold{src}. The image source.. Type: string. Default: NA.
\item \bold{srcSet}. The image srcSet.. Type: string. Default: NA.
\item \bold{sizes}. The image sizes.. Type: string. Default: NA.
\item \bold{alt}. The image alt.. Type: string. Default: NA.
\item \bold{width}. The image width.. Type: number. Default: NA.
\item \bold{height}. The image height.. Type: number. Default: NA.
\item \bold{radius}. The image border radius.. Type: none | sm | md | lg | full. Default: xl.
\item \bold{shadow}. The image shadow.. Type: none | sm | md | lg. Default: none.
\item \bold{loading}. A loading strategy to use for the image.. Type: eager | lazy. Default: NA.
\item \bold{fallbackSrc}. The fallback image source.. Type: string. Default: NA.
\item \bold{isBlurred}. Whether the image should have a duplicated blurred image at the background.. Type: boolean. Default: false.
\item \bold{isZoomed}. Whether the image should be zoomed when hovered.. Type: boolean. Default: false.
\item \bold{removeWrapper}. Whether to remove the wrapper element. This will cause the image to be rendered as a direct child of the parent element. If you set this prop as true neither the skeleton nor the zoom effect will work.. Type: boolean. Default: false.
\item \bold{disableSkeleton}. Whether the image should disable the skeleton animation while loading.. Type: boolean. Default: false.
\item \bold{classNames}. Allows to set custom class names for the image slots.. Type: Record<"img"｜ "wrapper"｜ "zoomedWrapper"｜ "blurredImg", string>. Default: NA.
}
\itemize{
\item \bold{onLoad}. Handler that is called when the image load.. Type: ReactEventHandler\if{html}{\out{<HTMLImageElement>}}. Default: NA.
\item \bold{onError}. Handler that is called when the image fails to load.. Type: () => void. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  div(
    class = "flex gap-5 my-2",
    image(
      width = 300,
      alt = "NextUI hero Image",
      src = "https://heroui.com/images/hero-card-complete.jpeg"
    ),
    image(
      isBlurred = TRUE,
      width = 300,
      alt = "Album cover",
      src = "https://heroui.com/images/hero-card-complete.jpeg"
    ),
    image(
      isZoomed = TRUE,
      width = 300,
      alt = "Fruit image with zoom effect",
      src = "https://heroui.com/images/hero-card-complete.jpeg"
    )
  )
)

server <- function(input, output, session) {
}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/image}.
}
