% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/inputs.R
\name{input}
\alias{input}
\alias{textInput}
\alias{numericInput}
\alias{dateInput}
\title{input}
\usage{
textInput(inputId, ..., value = defaultValue)

numericInput(inputId, ..., value = defaultValue)

dateInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{inputId}{ID of the component.}

\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{value}{Starting value.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Inputs fields are used to get the user input in a text field.
}
\details{
\enumerate{
\item Input Props
\itemize{
\item \bold{value}. number. Type: string. Default: NA.
\item \bold{initialValue}. Input default value. Type: string. Default: NA.
\item \bold{placeholder}. The short hint displayed in the input. Type: string. Default: NA.
\item \bold{size}. Change input size. Type: NormalSizes. Default: medium.
\item \bold{color}. Change input text, border and label color. Type: SimpleColors. Default: default.
\item \bold{status}. Change input status color. Type: SimpleColors. Default: default.
\item \bold{helperColor}. Change helper text color. Type: SimpleColors. Default: default.
\item \bold{required}. Required input prop. Type: boolean. Default: false.
\item \bold{readOnly}. It prevents the user from changing the value of the field. Type: boolean. Default: false.
\item \bold{disabled}. Disable input. Type: boolean. Default: false.
\item \bold{clearable}. Show clear button. Type: boolean. Default: false.
\item \bold{rounded}. Rounded input. Type: boolean. Default: false.
\item \bold{bordered}. Bordered input. Type: boolean. Default: false.
\item \bold{underlined}. Underlined input. Type: boolean. Default: false.
\item \bold{shadow}. Enable or disable the input shadow. Type: boolean. Default: true.
\item \bold{animated}. Enable or disable the input animation. Type: boolean. Default: true.
\item \bold{autoComplete}. HTML input autocomplete attribute. Type: string. Default: off.
\item \bold{borderWeight}. Border weight for bordered input. Type: NormalWeights. Default: normal.
\item \bold{fullWidth}. If true, the input will take up the full width of its container.. Type: boolean. Default: false.
\item \bold{width}. Input width. Type: string. Default: initial.
\item \bold{label}. Text label for input. Type: string. Default: NA.
\item \bold{labelPlaceholder}. The placeholder becomes a label. Type: string. Default: NA.
\item \bold{labelLeft}. Text label at left of the input. Type: string. Default: NA.
\item \bold{labelRight}. Text label at right of the input. Type: string. Default: NA.
\item \bold{helperText}. Add a helper text to Input. Type: string. Default: NA.
\item \bold{contentLeft}. Left content for input. Type: React.ReactNode. Default: NA.
\item \bold{contentRight}. Right content for input. Type: React.ReactNode. Default: NA.
\item \bold{contentClickable}. Left/right content are clickable (just applied when the styling prop is true, see the next props). Type: boolean. Default: false.
\item \bold{contentLeftStyling}. Allows the Input component to wrap the contentLeft in a container. Type: boolean. Default: true.
\item \bold{contentRightStyling}. Allows the Input component to wrap the contentRight in a container. Type: boolean. Default: true.
\item \bold{onChange}. Callback fired when the value is changed. Type: (e: React.ChangeEvent\if{html}{\out{<HTMLInputElement>}}) => void. Default: NA.
\item \bold{onFocus}. Callback fired when the input is focused.. Type: (e: React.FocusEvent\if{html}{\out{<HTMLInputElement>}}) => void. Default: NA.
\item \bold{onBlur}. Callback fired when the input is blurred.. Type: (e: React.FocusEvent\if{html}{\out{<HTMLInputElement>}}) => void. Default: NA.
\item \bold{onContentClick}. click icon event. Type: (key: ContentPosition, e: React.MouseEvent) => void. Default: NA.
\item \bold{onClearClick}. clear icon event. Type: (e: React.MouseEvent) => void. Default: NA.
\item \bold{ref}. forwardRef. Type: Ref<HTMLInputElement | null>. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: input.
\item \bold{...}. Input native props. Type: InputHTMLAttributes. Default: NA.
}
\item NA
\itemize{
\item \bold{hideToggle}. Hide toggle icon. Type: boolean. Default: false.
\item \bold{visibleIcon}. Custom visible password icon. Type: React.ReactNode. Default: NA.
\item \bold{hiddenIcon}. Custom hidden password icon. Type: React.ReactNode. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: input.
\item \bold{...}. Input props. Type: Input Props. Default: NA.
}
}
}
\seealso{
See \url{https://nextui.org/docs/components/input}.
}
