% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{link}
\alias{link}
\title{link}
\usage{
link(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Links allow users to click their way from page to page. This component is styled to resemble a hyperlink and semantically renders an \if{html}{\out{<a>}}
}
\details{
\enumerate{
\item Link Props
\itemize{
\item \bold{color}. Change link color. Type: LinkColors boolean string. Default: false.
\item \bold{href}. Link url. Type: string. Default: NA.
\item \bold{isExternal updated}. Show link icon. Type: boolean. Default: false.
\item \bold{underline}. Display underline. Type: boolean. Default: false.
\item \bold{block}. Display as a separate block. Type: boolean. Default: false.
\item \bold{ref}. forwardRef. Type: Ref<HTMLAnchorElement | null>. Default: NA.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: a.
\item \bold{...}. Native props. Type: AnchorHTMLAttributes. Default: NA.
}
}
}
\examples{
library(shiny)
library(shinyNextUI)

link <- shinyNextUI:::component("Link")
colors <- c(
  "text",
  "default",
  "primary",
  "secondary",
  "success",
  "warning",
  "error",
  "gradient"
)

link_config <- data.frame(
  color = colors,
  underline = c(rep(FALSE, 4), rep(TRUE, 4)),
  block = c(rep(TRUE, 4), rep(FALSE, 4)),
  href = c(rep("#", 4), rep("https://google.com", 4)),
  isExternal = c(rep(FALSE, 4), rep(TRUE, 4))
)

link_factory <- function(color, underline, block, href, isExternal) {
  grid(
    link(
      "A super link!",
      color = color,
      underline = underline,
      block = block,
      href = href,
      isExternal = isExternal
    )
  )
}

links <- purrr::pmap(link_config, link_factory)

ui <- nextui_page(
  text("Link"),
  grid_container(
    gap = 2,
    xs = 12,
    sm = 6,
    links
  )
)

server <- function(input, output, session) {}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/link}.
}
