% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{nextui_page}
\alias{nextui_page}
\title{NextUI page wrapper}
\usage{
nextui_page(..., dark_mode = FALSE, debug_react = FALSE)
}
\arguments{
\item{...}{UI elements.}

\item{dark_mode}{Apply global dark mode. If NULL, no switch is shown.}

\item{debug_react}{Whether to enable react debug mode.
Default to FALSE.}
}
\value{
Object which can be passed as the UI of a Shiny app.
}
\description{
Suppressed Bootstrap dependency which is not
needed.
}
