% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poke_data}
\alias{poke_data}
\title{Pokemon API data}
\format{
\subsection{\code{poke_data}}{

A nested list with 151 entries. Each sublist contains:
\describe{
\item{name}{Char: Pokemon name.}
\item{description}{Char: Pokemon description.}
\item{shape}{Char: Pokemon shape.}
\item{sprites}{List: Front and back sprites (images).}
\itemize{
\item{front_default}{Char: front sprite URL}
\item{shiny_default}{Char: front sprite URL (shiny form)}
}
\item{habitat}{Char: Pokemon habitat.}
...
}
}
}
\source{
\url{https://pokeapi.co/docs/v2}
}
\usage{
poke_data
}
\description{
Extract of some data from the 151 first Pokemons.
}
\note{
Have a look to inst/app-doc/data-doc.html to get an
interactive overview.
}
\keyword{datasets}
