% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/feedback.R
\name{popover}
\alias{popover}
\alias{popover_trigger}
\alias{popover_content}
\title{popover}
\usage{
popover(...)

popover_trigger(...)

popover_content(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Popover is a non-modal dialog that floats around its disclosure. It's commonly used for displaying additional rich content on top of something.
}
\details{
\itemize{
\item \bold{children*}. The content of the popover. It's usually the PopoverTrigger and PopoverContent.. Type: ReactNode[]. Default: NA.
\item \bold{size}. The popover content font size.. Type: sm | md | lg. Default: md.
\item \bold{color}. The popover color theme.. Type: default | primary | secondary | success | warning | danger. Default: default.
\item \bold{radius}. The popover border radius.. Type: none | sm | md | lg | full. Default: lg.
\item \bold{shadow}. The popover shadow.. Type: none | sm | md | lg. Default: lg.
\item \bold{backdrop}. The popover backdrop type.. Type: transparent | opaque | blur. Default: transparent.
\item \bold{placement}. The placement of the popover relative to its trigger reference.. Type: PopoverPlacement. Default: bottom.
\item \bold{state}. The controlled state of the popover. See Overlay States. Type: OverlayTriggerState. Default: NA.
\item \bold{isOpen}. Whether the popover is open by default (controlled).. Type: boolean. Default: NA.
\item \bold{defaultOpen}. Whether the popover is open by default (uncontrolled).. Type: boolean. Default: NA.
\item \bold{offset(px)}. The distance or margin between the reference and popper. It is used internally to create an offset modifier.. Type: number. Default: 7.
\item \bold{containerPadding(px)}. The placement padding that should be applied between the element and its surrounding container.. Type: number. Default: 12.
\item \bold{crossOffset(px)}. The additional offset applied along the cross axis between the element and its anchor element.. Type: number. Default: 0.
\item \bold{triggerType}. Type of popover that is opened by the trigger.. Type: dialog | menu | listbox | tree | grid;. Default: dialog.
\item \bold{showArrow}. Whether the popover should have an arrow.. Type: boolean. Default: false.
\item \bold{shouldFlip}. Whether the popover should change its placement and flip when it's about to overflow its boundary area.. Type: boolean. Default: true.
\item \bold{triggerScaleOnOpen}. Whether the trigger should scale down when the popover is open.. Type: boolean. Default: true.
\item \bold{shouldBlockScroll}. Whether to block scrolling outside the popover.. Type: boolean. Default: false.
\item \bold{isKeyboardDismissDisabled}. Whether pressing the escape key to close the popover should be disabled.. Type: boolean. Default: false.
\item \bold{shouldCloseOnBlur}. Whether the popover should close when focus is lost or moves outside it.. Type: boolean. Default: false.
\item \bold{motionProps}. The props to modify the framer motion animation. Use the variants API to create your own animation.. Type: MotionProps. Default: .
\item \bold{portalContainer}. The container element in which the overlay portal will be placed.. Type: HTMLElement. Default: document.body.
\item \bold{triggerRef}. The ref for the element which the popover positions itself with respect to.. Type: RefObject\if{html}{\out{<HTMLElement>}}. Default: NA.
\item \bold{scrollRef}. A ref for the scrollable region within the popover.. Type: RefObject\if{html}{\out{<HTMLElement>}}. Default: overlayRef.
\item \bold{disableAnimation}. Whether the popover is animated.. Type: boolean. Default: false.
\item \bold{classNames}. Allows to set custom class names for the popover slots.. Type: Record<"base"｜ "trigger"｜ "backdrop"｜ "content", string>. Default: NA.
}
\itemize{
\item \bold{onOpenChange}. Handler that is called when the popover's open state changes.. Type: (isOpen: boolean) => void. Default: NA.
\item \bold{shouldCloseOnInteractOutside}. When user interacts with the argument element outside of the popover ref, return true if onClose should be called. This gives you a chance to filter out interaction with elements that should not dismiss the popover. By default, onClose will always be called on interaction outside the overlay ref.. Type: (e: HTMLElement) => void. Default: NA.
\item \bold{onClose}. Handler that is called when the popover should close.. Type: () => void. Default: NA.
}
\itemize{
\item \bold{children*}. The popover trigger component, ensure the children passed is focusable. Users can tab to it using their keyboard, and it can take a ref. It is critical for accessiblity.. Type: ReactNode. Default: NA.
}
\itemize{
\item \bold{children}. The content that is displayed when the trigger is pressed.. Type: ReactNode. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  popover(
    showArrow = TRUE,
    placement = "right",
    popover_trigger(button("Click me!", color = "primary")),
    popover_content(
      "This is the content of the popover."
    )
  )
)

server <- function(input, output, session) {}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/popover}.
}
