% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{progress}
\alias{progress}
\title{progress}
\usage{
progress(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
The Progress component allows you to view the progress of any activity.
}
\details{
\itemize{
\item \bold{label}. The content to display as the label.. Type: ReactNode. Default: NA.
\item \bold{size}. The size of the indicator.. Type: sm | md | lg. Default: md.
\item \bold{color}. The color of the indicator.. Type: default | primary | secondary | success | warning | danger. Default: primary.
\item \bold{radius}. The progress border radius.. Type: none | sm | md | lg | full. Default: full.
\item \bold{value}. The current value (controlled).. Type: number. Default: NA.
\item \bold{valueLabel}. The content to display as the value's label (e.g. 1 of 4).. Type: ReactNode. Default: NA.
\item \bold{minValue}. The smallest value allowed for the input.. Type: number. Default: 0.
\item \bold{maxValue}. The largest value allowed for the input.. Type: number. Default: 100.
\item \bold{formatOptions}. The options to format the value.. Type: Intl.NumberFormat. Default: {style: 'percent'}.
\item \bold{isIndeterminate}. Whether the progress is indeterminate.. Type: boolean. Default: false.
\item \bold{isStriped}. Whether the progress is striped.. Type: boolean. Default: false.
\item \bold{showValueLabel}. Whether to show the value label.. Type: boolean. Default: true.
\item \bold{isDisabled}. Whether the progress is disabled.. Type: boolean. Default: false.
\item \bold{disableAnimation}. Whether to disable the animation.. Type: boolean. Default: false.
\item \bold{classNames}. Allows to set custom class names for the progress slots.. Type: Record<"base"｜"labelWrapper"｜"label"｜"track"｜"value"｜"track"｜"indicator", string>. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)

colors <- c(
  "default",
  "primary",
  "secondary",
  "success",
  "warning",
  "danger"
)

bool_par <- c(rep(FALSE, 3), rep(TRUE, 2), FALSE)

progress_config <- data.frame(
  label = c(rep("A label", 3), rep("", 3)),
  show_value_label = c(rep(FALSE, 3), rep(TRUE, 3)),
  size = c("sm", "sm", rep("md", 2), "lg", "lg"),
  color = colors,
  striped = bool_par,
  radius = c(rep("none", 2), "sm", "md", "lg", "full")
)

progress_factory <- function(
  label,
  show_value_label,
  size,
  color,
  striped,
  radius
) {
  progress(
    label = label,
    showValueLabel = show_value_label,
    value = round(runif(1, 0, 100)),
    size = size,
    color = color,
    isStriped = striped,
    radius = radius
  )
}

progresses <- purrr::pmap(progress_config, progress_factory)

ui <- nextui_page(
  div(
    class = "grid gap-4 grid-cols-3 grid-rows-3 m-5",
    progresses
  )
)

server <- function(input, output, session) {
}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/progress}.
}
