% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{radio_input}
\alias{radio_input}
\alias{update_radio_input}
\title{Radio input}
\usage{
radio_input(inputId, ..., choices, selected = choices[1])

update_radio_input(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...,
  choices = NULL,
  selected = NULL
)
}
\arguments{
\item{inputId}{Unique input id.}

\item{...}{Props.}

\item{choices}{Radio choices.}

\item{selected}{Default selected value.}

\item{session}{Shiny session.}
}
\value{
A radio input tag.
}
\description{
Radio input
}
