% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/inputs.R
\name{textarea}
\alias{textarea}
\alias{textarea_input}
\alias{update_textarea_input}
\title{textarea}
\usage{
textarea_input(inputId, ..., value = default_value)

update_textarea_input(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{inputId}{ID of the component.}

\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}

\item{value}{Starting value.}

\item{session}{Object passed as the \code{session} argument to Shiny server.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
The update functions return nothing (called for side effects).
}
\description{
Textarea component is a multi-line Input which allows you to write large texts.
}
\details{
\itemize{
\item \bold{children}. The content of the textarea.. Type: ReactNode. Default: NA.
\item \bold{minRows}. The minimum number of rows to display.. Type: number. Default: 3.
\item \bold{maxRows}. Maximum number of rows up to which the textarea can grow.. Type: number. Default: 8.
\item \bold{cacheMeasurements}. Reuse previously computed measurements when computing height of textarea.. Type: boolean. Default: false.
\item \bold{variant}. The variant of the textarea.. Type: flat | bordered | faded | underlined. Default: flat.
\item \bold{color}. The color of the textarea.. Type: default | primary | secondary | success | warning | danger. Default: default.
\item \bold{size}. The size of the textarea.. Type: sm|md|lg. Default: md.
\item \bold{radius}. The radius of the textarea.. Type: none | sm | md | lg | full. Default: NA.
\item \bold{label}. The content to display as the label.. Type: ReactNode. Default: NA.
\item \bold{value}. The current value of the textarea (controlled).. Type: string. Default: NA.
\item \bold{defaultValue}. The default value of the textarea (uncontrolled).. Type: string. Default: NA.
\item \bold{placeholder}. The placeholder of the textarea.. Type: string. Default: NA.
\item \bold{startContent}. Element to be rendered in the left side of the input.. Type: ReactNode. Default: NA.
\item \bold{endContent}. Element to be rendered in the right side of the input.. Type: ReactNode. Default: NA.
\item \bold{description}. A description for the textarea. Provides a hint such as specific requirements for what to choose.. Type: ReactNode. Default: NA.
\item \bold{errorMessage}. An error message for the textarea.. Type: ReactNode. Default: NA.
\item \bold{labelPlacement}. The position of the label.. Type: inside | outside | outside-left. Default: inside.
\item \bold{fullWidth}. Whether the textarea should take up the width of its parent.. Type: boolean. Default: true.
\item \bold{isRequired}. Whether user input is required on the textarea before form submission.. Type: boolean. Default: false.
\item \bold{isReadOnly}. Whether the textarea can be selected but not changed by the user.. Type: boolean. Default: .
\item \bold{isDisabled}. Whether the textarea is disabled.. Type: boolean. Default: false.
\item \bold{isInvalid}. Whether the textarea is invalid.. Type: boolean. Default: false.
\item \bold{validationState}. Whether the textarea should display its "valid" or "invalid" visual styling. (Deprecated) use isInvalid instead.. Type: valid | invalid. Default: NA.
\item \bold{disableAutosize}. Whether the textarea auto vertically resize should be disabled.. Type: boolean. Default: false.
\item \bold{disableAnimation}. Whether the textarea should be animated.. Type: boolean. Default: false.
\item \bold{classNames}. Allows to set custom class names for the checkbox slots.. Type: Record<"base"｜ "label"｜ "inputWrapper"｜ "innerWrapper" ｜ "input" ｜ "description" ｜ "errorMessage", string>. Default: NA.
}
\itemize{
\item \bold{onChange}. Handler that is called when the element's value changes. You can pull out the new value by accessing event.target.value (string).. Type: React.ChangeEvent \if{html}{\out{<HTMLInputElement>}}. Default: NA.
\item \bold{onValueChange}. Handler that is called when the element's value changes.. Type: (value: string) => void. Default: NA.
\item \bold{onClear}. Handler that is called when the clear button is clicked.. Type: () => void. Default: NA.
\item \bold{onHeightChange}. Handler that is called when the height of the textarea changes.. Type: (height: number, meta: { rowHeight: number }) => void. Default: NA.
}
}
\examples{
library(shiny)
library(shinyNextUI)

ui <- nextui_page(
  div(
    class = "flex gap-5",
    action_button("update_text", "Update text"),
    textarea_input(
      inputId = "textarea",
      placeholder = "Enter your amazing ideas.",
      label = "Text area input",
      bordered = TRUE,
      color = "secondary",
      status = "secondary",
      helperColor = "error",
      helperText = "Click on update text"
    )
  ),
  textOutput("textarea_val")
)

server <- function(input, output, session) {
  output$textarea_val <- renderText(input$textarea)

  observeEvent(input$update_text, {
    update_textarea_input(
      inputId = "textarea",
      value = "Updated value"
    )
  })
}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/textarea}.
}
