% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/doc.R
\name{user}
\alias{user}
\title{user}
\usage{
user(...)
}
\arguments{
\item{...}{Props to pass to the component.
The allowed props are listed below in the \bold{Details} section.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
Flexible User Profile Component.
}
\details{
\enumerate{
\item User Props
\itemize{
\item \bold{color}. Change User Profile border color. Type: NormalColors string. Default: default.
\item \bold{src}. Image source (local or remote). Type: string. Default: NA.
\item \bold{name}. Display Text Next to Image. Type: string. Default: NA.
\item \bold{text}. Display Text when image is missing. Type: string. Default: NA.
\item \bold{alt}. Display alt text when image is missing. Type: string. Default: NA.
\item \bold{size}. User Profile size. Type: NormalSizes number. Default: medium.
\item \bold{bordered}. Bordered User Profile. Type: boolean. Default: false.
\item \bold{zoomed}. Zoomed User Profile. Type: boolean. Default: false.
\item \bold{pointer}. Display pointer cursor on hover. Type: boolean. Default: false.
\item \bold{squared}. Squared User Profile. Type: boolean. Default: false.
\item \bold{css}. Override Default CSS style. Type: Stitches.CSS. Default: NA.
\item \bold{as}. Changes which tag component outputs. Type: keyof JSX.IntrinsicElements. Default: span.
\item \bold{...}. Native props. Type: ImgHTMLAttributes. Default: NA.
}
}
}
\examples{
library(shiny)
library(shinyNextUI)
library(shiny.react)

ui <- nextui_page(
  div(
    class = "grid gap-4 grid-cols-3 grid-rows-3 m-5",
    user(
      name = "Jane Doe",
      description = "Product Designer",
      avatarProps = JS("{
        src: 'https://i.pravatar.cc/150?u=a04258114e29026702d'
      }")
    )
  )
)

server <- function(input, output, session) {}

if (interactive() || is_testing()) shinyApp(ui, server)
}
\seealso{
See \url{https://nextui.org/docs/components/user}.
}
